fixed calories arithmetic error, dangling meals on regenerate plan
